refactor: bookWriter 설정 클래스 분리
